// app/interfaces/IReport.ts
export interface IReport {
    _id: any;
    site_id: any;
    absent_workers: any;
    late_workers: any;
    active_hours: any;
    inactive_hours: any;
  }

  export interface IReportInput{
    site_id: any;
    absent_workers: any;
    late_workers: any;
    active_hours: any;
    inactive_hours: any;
  }
  